♻️ [Refactoring]: api/dataモジュールをディレクトリ構造に再編成
- component、file、image-export、imageを独立したディレクトリへ移動
- インポートパスを新しいディレクトリ構造に合わせて更新
